medical_facts_evaluation/requirements.txt
# Medical Facts Evaluation
# Core dependencies

# HTTP client
requests>=2.28.0

# OpenAI for RAGAS
openai>=1.0.0

# RAGAS evaluation framework
ragas>=0.1.0

# Data processing
pandas>=2.0.0

# Beautiful terminal output
rich>=13.0.0

# Environment variable support
python-dotenv>=1.0.0

# Type hints (Python 3.9+)
typing-extensions>=4.0.0
